requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
blinker==1.7.0
brax==0.10.3
certifi==2024.2.2
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
cloudpickle==3.0.0
contextlib2==21.6.0
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
etils==1.8.0
Farama-Notifications==0.0.4
filelock==3.13.4
Flask==3.0.3
Flask-Cors==4.0.0
flax==0.8.2
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
glfw==2.7.0
grpcio==1.62.2
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
hydra-core==1.3.2
idna==3.7
imageio==2.34.1
importlib_resources==6.4.0
itsdangerous==2.2.0
jax==0.4.26
jaxlib==0.4.26
jaxopt==0.8.3
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.0.8
mujoco==3.1.4
mujoco-mjx==3.1.4
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opt-einsum==3.3.0
optax==0.2.2
orbax-checkpoint==0.5.10
packaging==24.0
pillow==10.3.0
protobuf==4.25.3
psutil==5.9.8
Pygments==2.17.2
PyOpenGL==3.1.7
pytinyrenderer==0.0.14
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
scipy==1.13.0
sentry-sdk==1.45.0
setproctitle==1.3.3
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
sympy==1.12
tensorboardX==2.6.2.2
tensorstore==0.1.56
toolz==0.12.1
torch==2.2.2
trimesh==4.3.1
typing_extensions==4.11.0
urllib3==2.2.1
wandb==0.16.6
Werkzeug==3.0.2
wheel==0.41.2
zipp==3.18.1